npm package
This fixes the build and adds an npm package. I had to add
LD_LIBRARY_PATH to fiber-shim because it fails to load if coroutine.so
is not in your current working directory. This also uses a better
heuristic for determining if fiber-shim was used.

Also the library is now just called `fibers` when used in the context of
node.